+++ app/src/client/components/cookie-consent/Banner.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import * as CookieConsent from 'vanilla-cookieconsent';
 import 'vanilla-cookieconsent/dist/cookieconsent.css';
-import getConfig from './CookieConsentConfig';
+import getConfig from './Config';
 
 /**
  * NOTE: if you do not want to use the cookie consent banner, you should first
