+++ app/src/client/components/cookie-consent/Config.ts
@@ -1,0 +1,117 @@
+import type { CookieConsentConfig } from 'vanilla-cookieconsent';
+
+declare global {
+  interface Window {
+    dataLayer: any;
+  }
+}
+
+const getConfig = () => {
+  
+  // See https://cookieconsent.orestbida.com/reference/configuration-reference.html for configuration options
+  const config: CookieConsentConfig = {
+    // Default configuration for the modal
+    root: 'body',
+    autoShow: true,
+    disablePageInteraction: true,
+    hideFromBots: true,
+    mode: 'opt-in',
+    revision: 0,
+
+    // Default configuration for the cookie
+    cookie: {
+      name: 'cc_cookie',
+      domain: location.hostname,
+      path: '/',
+      sameSite: 'Lax',
+      expiresAfterDays: 365,
+    },
+
+    guiOptions: {
+      consentModal: {
+        layout: 'box',
+        position: 'bottom right',
+        equalWeightButtons: true,
+        flipButtons: false,
+      },
+    },
+
+    categories: {
+      necessary: {
+        enabled: true, // this category is enabled by default
+        readOnly: true, // this category cannot be disabled
+      },
+      analytics: {
+        autoClear: {
+          cookies: [
+            {
+              name: /^_ga/, // regex: match all cookies starting with '_ga'
+            },
+            {
+              name: '_gid', // string: exact cookie name
+            },
+          ],
+        },
+
+        // a https://cookieconsent.orestbida.com/reference/configuration-reference.html#category-services
+        services: {
+          ga: {
+            label: 'Google Analytics',
+            onAccept: () => {
+              try {
+                // TODO: Make sure to replace this with your own GA ID, e.g. 'A-0123456789'
+                const GA_MEASUREMENT_ID = '';
+                if (!GA_MEASUREMENT_ID.length) {
+                  throw new Error('Google Analytics Measurement ID is missing');
+                }
+                window.dataLayer = window.dataLayer || [];
+                function gtag(..._args: unknown[]) {
+                  (window.dataLayer as Array<any>).push(arguments);
+                }
+                gtag('js', new Date());
+                gtag('config', GA_MEASUREMENT_ID);
+
+                // Adding the script tag dynamically to the DOM
+                const script = document.createElement('script');
+                script.src = `https://www.googletagmanager.com/gtag/js?id=${GA_MEASUREMENT_ID}`;
+                script.async = true;
+                document.body.appendChild(script);
+              } catch (error) {
+                console.error(error);
+              }
+            },
+            onReject: () => {},
+          },
+        },
+      },
+    },
+
+    language: {
+      default: 'en',
+      translations: {
+        en: {
+          consentModal: {
+            title: 'We use cookies',
+            description:
+              'We use cookies primarily for analytics to enhance your experience. By accepting, you agree to our use of these cookies. You can manage your preferences or learn more about our cookie policy.',
+            acceptAllBtn: 'Accept all',
+            acceptNecessaryBtn: 'Reject all',
+            // showPreferencesBtn: 'Manage Individual preferences',
+            footer: `
+            <a href="#" target="_blank">Privacy Policy</a>
+            <a href="#" target="_blank">Terms and Conditions</a>
+                    `,
+          },
+          // The showPreferencesBtn activates this modal to manage individual preferences https://cookieconsent.orestbida.com/reference/configuration-reference.html#translation-preferencesmodal
+          preferencesModal: {
+            sections: [],
+          },
+        },
+      },
+    },
+  };
+
+  return config;
+};
+
+export default getConfig;
